+++ test/FlxBasic.js
